+++ src/model/record-array.js
@@ -1,19 +1,24 @@
+import RESTless from '../core';
+import ModelStateMixin from './state';
+import Model from './model';
+
+var get = Ember.get;
+
 /**
   RecordArray is an Array of Model objects.
 
   @class RecordArray
   @namespace RESTless
   @extends Ember.ArrayProxy
-  @uses RESTless.State
 */
-RESTless.RecordArray = Ember.ArrayProxy.extend( RESTless.State, {
+var RecordArray = Ember.ArrayProxy.extend( ModelStateMixin, {
   /**
     The default adapter for the RecordArray. Providing a hook for overriding.
     @property adapter
    */
-  adapter: computed(function() {
+  adapter: Ember.computed('RESTless.client.adapter', function() {
     return get(RESTless, 'client.adapter');
-  }).property('RESTless.client.adapter'),
+  }),
 
   /**
     Use the current Serializer to turn the data into a record array.
@@ -23,8 +28,10 @@ RESTless.RecordArray = Ember.ArrayProxy.extend( RESTless.State, {
     @returns RESTless.RecordArray
    */
   deserializeMany: function(type, data) {
-    this._initContent();
     type = type || this.typeOfContent();
+    if(!this.content) { 
+      this.set('content', Ember.A());
+    }
     return get(this, 'adapter.serializer').deserializeMany(this, type, data);
   },
 
@@ -64,53 +71,51 @@ RESTless.RecordArray = Ember.ArrayProxy.extend( RESTless.State, {
   },
 
   /**
-    Helper to initialize the content property of the RecordArray if not present.
-    @private
-    @method _initContent
-    @returns RecordArray this
-   */
-  _initContent: function() {
-    if(!this.content) { 
-      this.set('content', Ember.A());
-    }
-    return this;
-  },
-
-  /**
     Observes when items become dirty and sets itself to dirty.
     @private
    */
-  _onItemDirtyChange: Ember.observer(function() {
-    var clean = this.get('content').everyBy('isDirty', false);
+  _onItemDirtyChange: Ember.observer('@each.isDirty', function() {
+    var clean = this.get('content').isEvery('isDirty', false);
     if(this.get('isLoaded') && !clean) {
       this.set('isDirty', true);
     }
-  }, '@each.isDirty'),
+  }),
 
   /**
     Observes when the array's isLoaded state changes and triggers each item's onLoaded.
     @private
    */
-  _onLoadedChange: Ember.observer(function() {
+  _onLoadedChange: Ember.observer('isLoaded', function() {
     if(this.get('isLoaded')) {
       this.forEach(function(item) {
-        if(RESTless.Model.detectInstance(item)) {
+        if(Model.detectInstance(item)) {
           item.onLoaded();
         }
       });
     }
-  }, 'isLoaded')
+  })
 });
 
 
-RESTless.RecordArray.reopenClass({
+RecordArray.reopenClass({
   /**
     Creates a RecordArray
     @method create
     @returns RESTless.RecordArray
    */
-  create: function() {
-    var arr = this._super.apply(this, arguments);
+  create: function(options) {
+    // If no content was provided, default to an empty array.
+    options = options || {};
+    if (!options.content) {
+      options.content = Ember.A();
+    }
+
+    // Properly apply rest args to super's create()
+    var restArgs = Array.prototype.slice.call(arguments, 1);
+    var args = [options].concat(restArgs);
+
+    var arr = this._super.apply(this, args);
+
     // override State defaults not implemented or applicable to arrays
     arr.setProperties({ _isReady: true, isNew: false });
     return arr;
@@ -121,7 +126,8 @@ RESTless.RecordArray.reopenClass({
     @returns RESTless.RecordArray
    */
   createWithContent: function() {
-    var arr = this.create.apply(this, arguments);
-    return arr._initContent();
+    return this.create.apply(this, arguments);
   }
 });
+
+export default RecordArray;