+++ src/transforms/boolean.js
@@ -1,4 +1,6 @@
-RESTless.BooleanTransform = RESTless.Transform.extend({
+import Transform from './base';
+
+export default Transform.extend({
   deserialize: function(serialized) {
     var type = typeof serialized;