+++ src/transforms/json.js
@@ -1,11 +1,16 @@
+import StringTransform from './string';
+import NumberTransform from './number';
+import BooleanTransform from './boolean';
+import DateTransform from './date';
+
 /**
   @property JSONTransforms
   @type Object
   @for RESTless
 */
-RESTless.JSONTransforms = {
-  'string'  : RESTless.StringTransform.create(),
-  'number'  : RESTless.NumberTransform.create(),
-  'boolean' : RESTless.BooleanTransform.create(),
-  'date'    : RESTless.DateTransform.create()
+export default {
+  'string'  : StringTransform.create(),
+  'number'  : NumberTransform.create(),
+  'boolean' : BooleanTransform.create(),
+  'date'    : DateTransform.create()
 };