+++ src/transforms/number.js
@@ -1,4 +1,8 @@
-RESTless.NumberTransform = RESTless.Transform.extend({
+import Transform from './base';
+
+var isEmpty = Ember.isEmpty;
+
+export default Transform.extend({
   deserialize: function(serialized) {
     return isEmpty(serialized) ? null : Number(serialized);
   },