+++ tests/pretender-server.js
@@ -1,0 +1,2 @@
+var server = new Pretender();
+server.unhandledRequest = function(){};